+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changes
 - 0.0.10
     - Sanity check duplicate mappings: Null entries do not lead to errors
+    - Updated README with the Java 64-bit requirement
 - 0.0.5
     - Fixed a bug where controller events are not properly picked up: Use controller event name, id and value instead of just controller event name to find mappings
     - Added a lot of event logging (for troubleshooting)